refactor: move SpreadsheetReferenceParser to correct package
